🐛 Fix mempool AlreadyOnTop issue, cleanup put() (#1414)
Issue: we stored a DP as AlreadyOnTop, but we then added the size of the
DP twice, modifying Fjall's metadata which is used for the querynewcut.
Then we committed a round with an incorrectly high lane-size, which we
then corrected, but the consensus failed in staking as a logic-issue,
and ends up being a bit stuck.

Fix is to hack everyone to agree on the higher value.